guardar por categorias

+++ view/producto/producto.php
@@ -1,14 +1,29 @@
 <main class="container-lg">
     <link rel="stylesheet" href="">
-    <h1 class="page-header" style="margin-bottom: 25px; margin-top: 25px;">Productos </h1>
-    <nav class="navbar navbar-light bg-light">
-        <div class="well well-sm text-right">
-            <a class="btn btn-primary" href="?c=proveedor">Ver Proveedores</a>
-            <a class="btn btn-primary" href="?c=producto&a=Nuevo">Nuevo Producto</a>
+    <center>
+        <h1 class="page-header" style="margin-bottom: 25px; margin-top: 25px;">ADMINISTRADOR </h1>
+    </center>
+    <div class="mb-5">
+        <div class="d-flex">
+            <div class="card me-5" style="width: 18rem;">
+                <div class="card-body">
+                    <h5 class="card-title">PROVEEDORES</h5>
+                    <p class="card-text">Espacio para ver y editar los proveedores registrados. Además de poder agregar proveedores nuevos.</p>
+                    <a class="btn btn-primary" href="?c=proveedor">Ver Proveedores</a>
+                </div>
+            </div>
+            <div class="card" style="width: 18rem;">
+                    <div class="card-body">
+                        <h5 class="card-title">PRODUCTOS</h5>
+                        <p class="card-text">Espacio para agregar productos nuevos.</p>
+                        <a class="btn btn-primary" href="?c=producto&a=Nuevo">Nuevo Producto</a>
+                    </div>
+                </div>
         </div>
-    </nav><br><br>
+    </div>
+    </div>
     <div class="container-producto container-fluid">
-        <?php foreach ($this->model->Listar() as $r): ?>
+        <?php foreach ($this->model->Listar() as $r) : ?>
             <div class="popo">
                 <div class="pipi">
                     <div class="producto_item">
@@ -42,15 +57,12 @@
                             </div>
                         </div>
                         <div>
-                        <td>
-                        <a href="?c=producto&a=Crud&id_art=<?php echo $r->id_art; ?>"><button type="button"
-                                class="btn btn-success"><i class="bi bi-pencil-fill"></i></button></a>
-                    </td>
-                    <td>
-                        <a onclick="javascript:return confirm('¿Seguro de eliminar este registro?');"
-                            href="?c=producto&a=Eliminar&id_art=<?php echo $r->id_art; ?>"><button type="button"
-                            class="btn btn-danger"><i class="bi bi-trash-fill"></i></button></a>
-                    </td>
+                            <td>
+                                <a href="?c=producto&a=Crud&id_art=<?php echo $r->id_art; ?>"><button type="button" class="btn btn-success"><i class="bi bi-pencil-fill"></i></button></a>
+                            </td>
+                            <td>
+                                <a onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=producto&a=Eliminar&id_art=<?php echo $r->id_art; ?>"><button type="button" class="btn btn-danger"><i class="bi bi-trash-fill"></i></button></a>
+                            </td>
                         </div><br>
                     </div>
                 </div>
